fix last config 0

+++ speedtest/src/main/scala/main/Program.scala
@@ -167,7 +167,7 @@ class SpeedTester(configs:sConfigParser, tablehandlers:Array[String], seed:Long)
           }
         } + " ")
         sb.append(sumOfUnfilteredCands/config.queriesSetSize) + " "
-        sb.append(sumOfUnfilteredCands/config.queriesSetSize/config.dataSetSize*100) + " "
+        sb.append(((sumOfUnfilteredCands.toFloat/config.queriesSetSize.toFloat)/config.dataSetSize.toFloat) *100) + " "
         sb.append(System.getProperty("line.separator"))
 
         // Write resulting set
